requirements.txt
openai
openai-whisper
python-dotenv
pyaudio

